add interface to Webmets, fix some bugs

+++ pymets/metric/length_metric.py
@@ -4,7 +4,8 @@ from pymets.model.swc_node import SwcTree
 from pymets.metric.utils.edge_match_utils import get_match_edges_e_fast
 from pymets.metric.utils.config_utils import get_default_threshold
 from pymets.io.read_json import read_json
-from pymets.io.save_swc import save_as_swc,print_swc
+from pymets.io.save_swc import save_as_swc, swc_to_list
+from pymets.io.read_swc import adjust_swcfile
 
 import time
 import os,platform
@@ -14,8 +15,8 @@ def length_metric_2(gold_swc_tree=None, test_swc_tree=None, dis_threshold=0.1, d
     test_swc_tree.get_lca_preprocess()
     match_edges, un_match_edges = get_match_edges_e_fast(gold_swc_tree, test_swc_tree,  # tree data
                                                          dis_threshold, detail_path, DEBUG=False)  # configs
-
-    save_as_swc(object=un_match_edges, file_path=detail_path)
+    if detail_path is not None:
+        save_as_swc(object=un_match_edges, file_path=detail_path)
     match_length = 0.0
     for line_tuple in match_edges:
         match_length += line_tuple[0].parent_distance()
@@ -41,6 +42,10 @@ def length_metric_1(gold_swc_tree=None, test_swc_tree=None, DEBUG=False):
 
 
 def length_metric(gold_swc_tree, test_swc_tree, abs_dir, config):
+    # remove old point mark
+    gold_swc_tree.type_clear(5)
+    test_swc_tree.type_clear(4)
+
     # get config threshold
     if "rad_threshold" not in config.keys() or config["rad_threshold"] == "default":
         rad_threshold = get_default_threshold(gold_swc_tree)
@@ -50,40 +55,74 @@ def length_metric(gold_swc_tree, test_swc_tree, abs_dir, config):
         except:
             raise Exception("[Error: ] Read config info threshold {}. suppose to be a float or \"default\"")
     # get config detail path
-    detail_path = config["detail"]
-    if platform.system() == "Linux":
-        detail_path = '/'.join(detail_path.split("\\"))
-    detail_path = os.path.join(abs_dir, detail_path)
-    if os.path.exists(detail_path):
-        os.remove(detail_path)
+    if "detail" in config.keys():
+        detail_path = config["detail"]
+        if platform.system() == "Linux":
+            detail_path = '/'.join(detail_path.split("\\"))
+        detail_path = os.path.join(abs_dir, detail_path)
+        if os.path.exists(detail_path):
+            os.remove(detail_path)
+    else:
+        detail_path = None
+
     # get config method
     if config["method"] == 1:
         ratio = length_metric_1(gold_swc_tree=gold_swc_tree,
                                 test_swc_tree=test_swc_tree)
         print("1 - test_length / gold_length= {}".format(ratio))
+        return ratio
     elif config["method"] == 2:
         recall, precision = length_metric_2(gold_swc_tree=gold_swc_tree,
                                             test_swc_tree=test_swc_tree,
                                             dis_threshold=rad_threshold,
                                             detail_path=detail_path,
-                                            DEBUG=True)
+                                            DEBUG=False)
         print("Recall = {}, Precision = {}".format(recall, precision))
+        return recall, precision
     else:
-        raise Exception("[Error: ] Read config info method {}. length metric only have 1 and 2 two methods")
+        raise Exception("[Error: ] Read config info method {}. length metric only have 1 and 2 two methods".format(
+            config["method"]
+        ))
+
+
+# length metric interface connect to webmets
+def web_length_metric(gold_swc, test_swc, method, rad_threshold, len_threshold):
+    gold_tree = SwcTree()
+    test_tree = SwcTree()
+
+    gold_tree.load_list(adjust_swcfile(gold_swc))
+    test_tree.load_list(adjust_swcfile(test_swc))
+
+
+    config = {
+        'method': method,
+        'length_threshold': len_threshold,
+        'rad_threshold': rad_threshold
+    }
+    print("method = {}".format(
+        config['method']
+    ))
+    recall, precision = length_metric(gold_swc_tree=gold_tree,
+                                      test_swc_tree=test_tree,
+                                      abs_dir="",
+                                      config=config)
+    gold_tree.radius_limit(10)
+    test_tree.radius_limit(10)
+    return recall, precision, swc_to_list(gold_tree), swc_to_list(test_tree)
 
 
 if __name__ == "__main__":
     goldtree = SwcTree()
 
     testTree = SwcTree()
-    goldtree.load("D:\gitProject\mine\PyMets\\test\data_example\gold\\2_18_gold.swc")
-    testTree.load("D:\gitProject\mine\PyMets\\test\data_example\\test\\2_18_test.swc")
+    goldtree.load("D:\gitProject\mine\PyMets\\test\data_example\gold\\34_23_10_gold.swc")
+    testTree.load("D:\gitProject\mine\PyMets\\test\data_example\\test\\34_23_10_test.swc")
 
     start = time.time()
-    # length_metric(gold_swc_tree=goldtree,
-    #               test_swc_tree=testTree,
-    #               abs_dir="D:\gitProject\mine\PyMets",
-    #               config=read_json("D:\gitProject\mine\PyMets\config\length_metric.json"))
+    length_metric(gold_swc_tree=goldtree,
+                  test_swc_tree=testTree,
+                  abs_dir="D:\gitProject\mine\PyMets",
+                  config=read_json("D:\gitProject\mine\PyMets\config\length_metric.json"))
     length_metric(gold_swc_tree=testTree,
                   test_swc_tree=goldtree,
                   abs_dir="D:\gitProject\mine\PyMets",
